Refactor test suite into focused, organized test classes
