+++ ImGuiApp.Test/ImGuiControllerComponentTests.cs
@@ -6,13 +6,19 @@ namespace ktsu.ImGuiApp.Test;
 
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using ktsu.ImGuiApp.ImGuiController;
+using Hexa.NET.ImGui;
 using System;
 using Moq;
 using Silk.NET.Windowing;
 
+/// <summary>
+/// Tests for ImGuiController components including font configuration, factories, and providers.
+/// </summary>
 [TestClass]
-public class ImGuiControllerTests
+public class ImGuiControllerComponentTests
 {
+	#region ImGuiFontConfig Tests
+
 	[TestMethod]
 	public void ImGuiFontConfig_Constructor_ValidParameters_InitializesCorrectly()
 	{
@@ -31,7 +37,7 @@ public class ImGuiControllerTests
 	{
 		const string testPath = "test.ttf";
 		const int testSize = 16;
-		static IntPtr TestGlyphRange(Hexa.NET.ImGui.ImGuiIOPtr io) => IntPtr.Zero;
+		static IntPtr TestGlyphRange(ImGuiIOPtr io) => IntPtr.Zero;
 
 		ImGuiFontConfig config = new(testPath, testSize, TestGlyphRange);
 
@@ -108,12 +114,31 @@ public class ImGuiControllerTests
 	}
 
 	[TestMethod]
+	public void ImGuiFontConfig_IsStruct()
+	{
+		Type configType = typeof(ImGuiFontConfig);
+		Assert.IsTrue(configType.IsValueType);
+		Assert.IsFalse(configType.IsClass);
+	}
+
+	[TestMethod]
+	public void ImGuiFontConfig_ImplementsIEquatable()
+	{
+		Type configType = typeof(ImGuiFontConfig);
+		Type equatableType = typeof(IEquatable<ImGuiFontConfig>);
+		Assert.IsTrue(equatableType.IsAssignableFrom(configType));
+	}
+
+	#endregion
+
+	#region WindowOpenGLFactory Tests
+
+	[TestMethod]
 	public void WindowOpenGLFactory_Constructor_ValidWindow_InitializesCorrectly()
 	{
 		Mock<IWindow> mockWindow = new();
 		WindowOpenGLFactory factory = new(mockWindow.Object);
 
-		// If constructor doesn't throw, the factory was created successfully
 		Assert.IsNotNull(factory);
 	}
 
@@ -123,6 +148,17 @@ public class ImGuiControllerTests
 	{
 		_ = new WindowOpenGLFactory(null!);
 	}
+
+	[TestMethod]
+	public void WindowOpenGLFactory_ImplementsIOpenGLFactory()
+	{
+		Type factoryType = typeof(WindowOpenGLFactory);
+		Assert.IsTrue(typeof(IOpenGLFactory).IsAssignableFrom(factoryType));
+	}
+
+	#endregion
+
+	#region OpenGLProvider Tests
 
 	[TestMethod]
 	public void OpenGLProvider_Constructor_ValidFactory_InitializesCorrectly()
@@ -153,7 +189,6 @@ public class ImGuiControllerTests
 		Mock<IOpenGLFactory> mockFactory = new();
 		OpenGLProvider provider = new(mockFactory.Object);
 
-		// Should not throw when called multiple times
 		provider.Dispose();
 		provider.Dispose();
 		Assert.IsTrue(true); // If we get here, no exception was thrown
@@ -169,4 +204,54 @@ public class ImGuiControllerTests
 		provider.Dispose();
 		_ = provider.GetGL(); // Should throw ObjectDisposedException
 	}
+
+	[TestMethod]
+	public void OpenGLProvider_ImplementsIDisposable()
+	{
+		Type providerType = typeof(OpenGLProvider);
+		Assert.IsTrue(typeof(IDisposable).IsAssignableFrom(providerType));
+	}
+
+	#endregion
+
+	#region Texture and TextureCoordinate Tests
+
+	[TestMethod]
+	public void TextureCoordinate_EnumValues_HaveCorrectMappings()
+	{
+		Assert.AreEqual(0, (int)TextureCoordinate.None);
+		Assert.AreEqual((int)Silk.NET.OpenGL.TextureParameterName.TextureWrapS, (int)TextureCoordinate.S);
+		Assert.AreEqual((int)Silk.NET.OpenGL.TextureParameterName.TextureWrapT, (int)TextureCoordinate.T);
+		Assert.AreEqual((int)Silk.NET.OpenGL.TextureParameterName.TextureWrapR, (int)TextureCoordinate.R);
+	}
+
+	[TestMethod]
+	public void TextureCoordinate_AllEnumValues_AreDefined()
+	{
+		TextureCoordinate[] values = Enum.GetValues<TextureCoordinate>();
+
+		Assert.IsTrue(values.Length >= 4);
+		Assert.IsTrue(values.Contains(TextureCoordinate.None));
+		Assert.IsTrue(values.Contains(TextureCoordinate.S));
+		Assert.IsTrue(values.Contains(TextureCoordinate.T));
+		Assert.IsTrue(values.Contains(TextureCoordinate.R));
+	}
+
+	[TestMethod]
+	public void TextureCoordinate_IsEnum()
+	{
+		Type coordType = typeof(TextureCoordinate);
+		Assert.IsTrue(coordType.IsEnum);
+		Assert.AreEqual(typeof(int), Enum.GetUnderlyingType(coordType));
+	}
+
+	[TestMethod]
+	public void Texture_Constants_HaveExpectedValues()
+	{
+		Assert.AreEqual((Silk.NET.OpenGL.SizedInternalFormat)Silk.NET.OpenGL.GLEnum.Srgb8Alpha8, Texture.Srgb8Alpha8);
+		Assert.AreEqual((Silk.NET.OpenGL.SizedInternalFormat)Silk.NET.OpenGL.GLEnum.Rgb32f, Texture.Rgb32F);
+		Assert.AreEqual((Silk.NET.OpenGL.GLEnum)0x84FF, Texture.MaxTextureMaxAnisotropy);
+	}
+
+	#endregion
 }
